Refactoring, moved Page class
