+++ src/test/groovy/com/ofg/geb/AcceptanceSwaggerUISpec.groovy
@@ -1,4 +1,6 @@
 package com.ofg.geb
+
+import com.ofg.geb.pages.SwaggerUIHomePage
 import geb.spock.GebSpec
 
 class AcceptanceSwaggerUISpec extends GebSpec{
