+++ src/test/groovy/com/ofg/geb/HealthEndpointUISpec.groovy
@@ -1,5 +1,6 @@
 package com.ofg.geb
 
+import com.ofg.geb.pages.SwaggerUIHomePage
 import geb.spock.GebSpec
 import spock.lang.Unroll
